requirements.txt
pybuilder
selenium 
flask
pillow
pandas
sklearn
graphviz
image
flask_mysqldb
